refactor: deduplicate test helpers into testutil_test.go
Extract initTestRepo, runGit, and writeFile from git_test.go into a
shared testutil_test.go. Remove the identical copies (initTestRepoForConfig,
runGitForConfig, writeFileForConfig) from config_test.go and update all
call sites to use the shared helpers.
